Add client receive
