+++ src/client.rs
@@ -1,4 +1,5 @@
 use std::future::{Future, IntoFuture};
+use std::io::{Bytes, Error};
 use futures_util::stream::{SplitSink, SplitStream};
 use tokio::sync::{broadcast, oneshot};
 use tokio_tungstenite::{connect_async, tungstenite, MaybeTlsStream, WebSocketStream};
@@ -9,34 +10,8 @@ use tokio::net::TcpStream;
 use tokio_tungstenite::tungstenite::handshake::client::Request;
 use crate::MessageG2V;
 use crate::schema::MessageV2G;
+use crate::stream::BlimpGroundWebsocketStreamPair;
 
-struct BlimpGroundWebsocketStreamPair {
-    read_stream: SplitStream<WebSocketStream<MaybeTlsStream<TcpStream>>>,
-    write_stream: SplitSink<WebSocketStream<MaybeTlsStream<TcpStream>>, tungstenite::Message>
-}
-
-impl BlimpGroundWebsocketStreamPair {
-    fn from_stream(stream: WebSocketStream<MaybeTlsStream<TcpStream>>) -> Self {
-        let (write_stream, read_stream) = stream.split();
-        Self {read_stream, write_stream}
-    }
-    async fn send(&mut self, message: impl Serialize) -> Result<(), Box<dyn std::error::Error>> {
-        let serialized = postcard::to_stdvec(&message)?;
-        self.write_stream.send(
-            tungstenite::Message::Binary(tungstenite::Bytes::from(serialized))
-        ).await?;
-        Ok(())
-    }
-
-    async fn recv(&mut self) -> Result<MessageG2V, Box<dyn std::error::Error>> {
-        todo!();
-    }
-
-
-    async fn close(&mut self) -> Result<(), tungstenite::Error> {
-        self.write_stream.close().await
-    }
-}
 pub struct BlimpGroundWebsocketClient {
     url: String,
     stream: Option<BlimpGroundWebsocketStreamPair>,
@@ -61,5 +36,9 @@ impl BlimpGroundWebsocketClient {
     }
     pub async fn send(&mut self, message: MessageV2G) -> Result<(), Box<dyn std::error::Error>> {
         self.stream.as_mut().unwrap().send(message).await
+    }
+
+    pub async fn recv(&mut self) -> Result<MessageG2V, Box<dyn std::error::Error>> {
+        self.stream.as_mut().unwrap().recv().await
     }
 }
