+++ src/lib.rs
@@ -1,5 +1,6 @@
 mod client;
 mod schema;
+mod stream;
 
 pub use client::BlimpGroundWebsocketClient;
 pub use schema::*;
